Retrieving client json data from API json_output_string
This should allow you to pipe the output of a script using this library to a file or something else. At the moment only implemented for client connections. Server instances using the same tactic are generating a segmentation fault. Have to see if this is a problem with the python wrapper or the iperf3 C libary itself